refactor(contacts): three desktop panes (#166)
BooksSidebar, ContactListPanel and ContactDetailPanel per
docs/features.md. Pane roots keep their classes so the view's
layout-mode rules still size them; interior CSS moves with each
pane. The list panel resolves the merge keeper/loser pair itself
and emits it; encodeMailtoAddress/sanitizeTel move to lib/mailto.
The view keeps data ownership, the seq-guarded fetchBooks +
contacts-changed listener, watchers, selection, and delete/merge
orchestration. +13 tests incl. the first ContactsView integration
coverage.
